requirements.txt
tqdm == 4.19.4
numpy == 1.14.0
docopt == 0.6.2
Pillow == 4.3.0
scikit_image == 0.13.1
guided_filter_pytorch == 1.0